SocialMediaManager/requirements.txt
agency-swarm==0.4.4
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.3
audioop-lts==0.2.1
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
datamodel-code-generator==0.26.1
deepdiff==6.7.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
fsspec==2024.12.0
genson==1.3.0
gradio==5.12.0
gradio_client==1.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
inflect==5.6.2
isort==5.13.2
Jinja2==3.1.5
jiter==0.8.2
jsonref==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==2.2.2
openai==1.59.9
ordered-set==4.1.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.7.1
ruff==0.9.2
safehttpx==0.1.6
safetensors==0.5.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
termcolor==2.4.0
tokenizers==0.21.0
tomlkit==0.13.2
tqdm==4.67.1
transformers==4.48.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
